Map.get returns Optional.none() if value is null.
I am not entirely sure if null values should be allowed in Maps in the first place, since it breaks the sequence abstraction - sequences require that all of their values are non-null. However, fact of the matter is that Maps currently permit null values to be put in them. This change allows us to get the values back out, as Optional.none()s.